requirements.txt
itsdangerous
Jinja2
MarkupSafe
Pillow
SQLAlchemy
virtualenv
Werkzeug
psycopg2
gunicorn
alembic
bleach
blinker
Click
psycopg2
python-dotenv
Flask
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
Flask-SeaSurf
Flask-Cors
validator-collection
